+++ packages/react/src/components/elements/index.ts
@@ -9,3 +9,4 @@ export * from "./attachment";
 export * from "./repeaterGroup";
 export * from "./multiSelect";
 export * from "./emailInput";
+export * from "./dateRangePicker";