+++ packages/types/src/elements/dateRangePicker/index.ts
@@ -1,0 +1,20 @@
+import {
+  GenericFieldAttributes,
+  GenericLezzformElement,
+  LezzformElementCategory,
+} from "../..";
+import { ObjectRule } from "../../rules";
+
+export type DateRangePickerType = GenericLezzformElement<
+  GenericFieldAttributes & DateRangePickerAttributesType
+> & {
+  type: "DateRangePicker";
+  category: LezzformElementCategory.Field;
+  rule: DateRangePickerRuleType;
+};
+
+export type DateRangePickerAttributesType = {
+  format: string;
+};
+
+export type DateRangePickerRuleType = ObjectRule;